# You-book Project/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "유북 API 명세서",
    "description": "Node.js Swaager swagger-jsdoc 방식 RestFul API 클라이언트 UI"
  },
  "servers": [
    {
      "url": "http://localhost:5001"
    }
  ],
  "paths": {
...
    "/get_category": {
      "get": {
        "summary": "사용자의 카테고리 목록 가져오기",
        "tags": [
          "Category"
        ],
        "responses": {
          "200": {
            "description": "사용자의 카테고리 목록을 반환",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "categorys": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "로그인하지 않은 사용자"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
    "/update_category": {
      "post": {
        "summary": "문단의 카테고리를 업데이트",
        "tags": [
          "Content"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "현재 카테고리 이름"
                  },
                  "new_name": {
                    "type": "string",
                    "description": "변경할 새로운 카테고리 이름"
                  },
                  "bookId": {
                    "type": "string",
                    "description": "책 ID"
                  },
                  "inputCount": {
                    "type": "integer",
                    "description": "입력 횟수"
                  },
                  "content_order": {
                    "type": "integer",
                    "description": "업데이트할 레코드의 content_order"
                  },
                  "category": {
                    "type": "string",
                    "description": "새롭게 업데이트할 카테고리 내용"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "카테고리가 성공적으로 업데이트됨",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "요청 데이터가 잘못됨"
          },
          "401": {
            "description": "로그인하지 않은 사용자"
          },
          "404": {
            "description": "해당하는 레코드가 없음"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
    "/delete_category": {
      "post": {
        "summary": "사용자의 카테고리 삭제",
        "tags": [
          "Category"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "삭제할 카테고리 이름"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "카테고리가 성공적으로 삭제됨",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "로그인하지 않은 사용자"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
    "/add_category": {
      "post": {
        "summary": "새로운 카테고리 추가",
        "tags": [
          "Category"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "추가할 카테고리 이름"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "카테고리가 성공적으로 추가됨",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "로그인하지 않은 사용자"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
...
    "/get_id": {
      "post": {
        "summary": "사용자 ID 정보 가져오기",
        "tags": [
          "User"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "사용자 이름"
                  },
                  "phoneNum": {
                    "type": "string",
                    "description": "사용자 전화번호"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "사용자 ID 반환",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "userId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "입력 값 누락"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
    "/confirm_user": {
      "post": {
        "summary": "사용자 존재 여부 확인",
        "tags": [
          "User"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "description": "사용자 ID"
                  },
                  "name": {
                    "type": "string",
                    "description": "사용자 이름"
                  },
                  "phoneNum": {
                    "type": "string",
                    "description": "사용자 전화번호"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "사용자 정보 반환",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "user": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "입력 값 누락"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
    "/reset_pw": {
      "post": {
        "summary": "사용자 비밀번호 재설정",
        "tags": [
          "User"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "description": "사용자 ID"
                  },
                  "name": {
                    "type": "string",
                    "description": "사용자 이름"
                  },
                  "phoneNum": {
                    "type": "string",
                    "description": "사용자 전화번호"
                  },
                  "newPW": {
                    "type": "string",
                    "description": "새 비밀번호"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "비밀번호 재설정 성공 여부 반환",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "userId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "입력 값 누락"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
...
  },
...
}